addressed code review changes, moved product API to a new file and added input validations
